+++ tai/__init__.py
@@ -19,10 +19,8 @@
 """Package initialization file."""
 
 # Python 3 compatibility
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
 
 # import io  # Python 3 compatibility
 import os