+++ tai/appinfo.py
@@ -19,10 +19,8 @@
 """Application basic information."""
 
 # Python 3 compatibility
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
 
 import datetime as dt
 # import io  # Python 3 compatibility
@@ -31,7 +29,7 @@ import datetime as dt
 
 
 APP_NAME = 'tai'
-APP_VERSION = '0.0.30'
+APP_VERSION = '0.0.31'
 APP_LICENSE = 'GNU General Public License v2 or later (GPLv2+)'
 APP_AUTHOR = 'Joao Carlos Roseta Matos'
 APP_EMAIL = 'jcrmatos@gmail.com'